Rename tools to toolbar_mode
