+++ glue/viewers/image/qt/contrast_mouse_mode.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import, division, print_function
 
 from glue.external.echo import delay_callback
 from glue.config import viewer_tool
-from glue.viewers.matplotlib.tools import ToolbarModeBase
+from glue.viewers.matplotlib.toolbar_mode import ToolbarModeBase
 
 
 @viewer_tool
