+++ glue/viewers/matplotlib/__init__.py
@@ -1,1 +1,1 @@
-from . import tools  # noqa
+from . import toolbar_mode  # noqa
